// tarea_4/src/webparts/tarea4/models/Grupos.ts
export default class Grupos {
    ID: number;
    SectorAsociado: any;
    SectorAsociadoId: string | number;
    CodigoDeGrupo: string;
    Denominacion: string;
    Descripcion: string;
    FechaDeCreacion: string;
    FechaDeFinalizacion: string;
    Estado: boolean;
    TipoDeGrupo: string;
    Tematica: string;
    // Pais: string;
    // Ciudad: string;
    // Ambito: string;
}