// tarea_4/src/webparts/tarea4/models/FormFields.ts
// import { IPickerTerms } from "@pnp/spfx-controls-react";

export default class FormFields {
    SectorAsociadoId: string | number;
    CodigoDeGrupo: string;
    Denominacion: string;
    Descripcion: string;
    FechaDeCreacion: Date;
    FechaDeFinalizacion: Date;
    Estado: boolean;
    TipoDeGrupo: string;
    Tematica: string;
    // Pais: IPickerTerms;
    // Ciudad: IPickerTerms;
    // Ambito: IPickerTerms;
}